Fixed chapter1 exercises
