+++ Chapter01/Exercises/Exercise109/Program.cs
@@ -1,38 +1,95 @@
 ﻿using System;
 
-class Program
+namespace Exercise108
 {
-    static void Main(string[] args)
+    class Program
     {
-        int[] randomNumbers = { 123, 22, 53, 91, 787, 0, -23, 5 };
-
-        int[] sortedArray = BubbleSort(randomNumbers);
-
-        Console.WriteLine("Sorted:");
-
-        for (int i = 0; i < sortedArray.Length; i++)
-            Console.Write(sortedArray[i] + " ");
-
-        Console.Read();
-    }
-
-    static int[] BubbleSort(int[] array)
-    {
-        int temp;
-
-        for (int j = 0; j < array.Length - 1; j++)
+        static void Main(string[] args)
         {
-            for (int i = 0; i < array.Length - 1; i++)
-            {
-                if (array[i] > array[i + 1])
-                {
-                    temp = array[i + 1];
-                    array[i + 1] = array[i];
-                    array[i] = temp;
-                }
-            }
+            Console.Write("Enter the Number to check Prime: ");
+
+            var input = int.Parse(Console.ReadLine());
+
+            Console.WriteLine($"{input} is prime? {IsPrimeWithGoTo(input)}.");
         }
 
-        return array;
+        static bool IsPrime(int number)
+        {
+            bool isPrime = true;
+
+            int counter = 2;
+
+            while (counter <= number / 2)
+            {
+                if (number % counter == 0)
+                {
+                    isPrime = false;
+                    break;
+                }
+
+                counter++;
+            }
+
+            return isPrime;
+        }
+
+        static bool IsPrimeWithContinue(int number)
+        {
+            bool isPrime = true;
+
+            int counter = 2;
+
+            while (counter <= number / 2)
+            {
+                if (number % counter != 0)
+                {
+                    counter++;
+                    continue;
+                }
+
+                isPrime = false;
+                break;
+            }
+
+            return isPrime;
+        }
+
+        static bool IsPrimeWithGoTo(int number)
+        {
+            bool isPrime = true;
+
+            int counter = 2;
+
+            while (counter <= number / 2)
+            {
+                if (number % counter == 0)
+                {
+                    isPrime = false;
+                    goto isNotAPrime; 
+                }
+
+                counter++;
+            }
+
+            isNotAPrime:
+            return isPrime;
+        }
+
+        static bool IsPrimeWithReturn(int number)
+        {
+            int counter = 2;
+
+            while (counter <= number / 2)
+            {
+                if (number % counter == 0)
+                {
+                    return false;
+                }
+
+                counter ++;
+            }
+
+            return true;
+        }
     }
 }
